Version 0.1!
- Updated README.md (now with installation tutorial!)
- Support for single-shot EPI readout events, which brings us to
- Spin Echo EPI implementation either using SpinEchoEpi_GPI Node or the conventonal way
- JEMRIS support
- Bug fixes

+++ mr_gpi/makedelay.py
@@ -7,7 +7,7 @@ def makedelay(d):
 
     Parameters
     ----------
-    d : int
+    d : float
         Delay time, in seconds.
 
     Returns
